requirements.txt
Django==4.0.6
django-durationwidget==1.0.5
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2-binary==2.9.3
requests==2.28.1
debugpy==1.6.3
python-dotenv==0.20.0
pytest-django==4.5.2
requests_mock